+++ packages/interactive_media_ads/pigeons/interactive_media_ads_android.dart
@@ -2,9 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(bparrishMines): Uncomment this file once
-// https://github.com/flutter/packages/pull/6371 lands. This file uses the
-// Kotlin ProxyApi feature from pigeon.
 // ignore_for_file: avoid_unused_constructor_parameters
 
 import 'package:pigeon/pigeon.dart';
@@ -327,7 +324,15 @@ abstract class AdsRequest {
         'com.google.ads.interactivemedia.v3.api.player.ContentProgressProvider',
   ),
 )
-abstract class ContentProgressProvider {}
+abstract class ContentProgressProvider {
+  ContentProgressProvider();
+
+  /// Sets an update on the progress of the video.
+  ///
+  /// This is a custom method added to the native class because the native
+  /// method `getContentProgress` requires a synchronous return value.
+  void setContentProgress(VideoProgressUpdate update);
+}
 
 /// An object which handles playing ads after they've been received from the
 /// server.