+++ packages/web_benchmarks/testing/test_app/lib/about_page.dart
@@ -4,8 +4,6 @@
 
 import 'package:flutter/material.dart';
 
-const ValueKey<String> backKey = ValueKey<String>('backKey');
-
 class AboutPage extends StatelessWidget {
   const AboutPage({super.key});
 
@@ -14,7 +12,6 @@ class AboutPage extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         leading: BackButton(
-          key: backKey,
           onPressed: () => Navigator.of(context).pop(),
         ),
       ),