+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.2
+
+* Adds support for mid-roll ads. See `AdsRequest.contentProgressProvider`.
+
+## 0.2.1
+
+* Adds internal wrapper for Android native `ContentProgressProvider`.
+
 ## 0.2.0
 
 * Adds support for pausing and resuming Ad playback. See `AdsManager.pause` and `AdsManager.resume`.